requirements.txt
boto3==1.9.183
botocore==1.12.183
certifi==2019.6.16
chardet==3.0.4
docutils==0.14
elasticsearch==7.0.2
entrypoints==0.3
flake8==3.7.8
idna==2.8
jmespath==0.9.4
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
requests-aws4auth==0.9
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
